Move blur metric function to skimage.measure
